showcase

+++ include/Interface/memory_manager_priv.h
@@ -3,5 +3,5 @@
 
 #include <stddef.h>
 void* mm_request_region(size_t size);
-void* mm_resize_region(void* old_ptr, size_t old_size, size_t new_size, uint32_t max_allowed_size);
+void* mm_resize_region(void* old_ptr, uint32_t old_size, uint32_t new_size, uint32_t max_allowed_size);
 #endif
